Refactor sort to deduplicate code
